move experiment controller logic to appropriate file
