+++ app/routes/experiments/info.js
@@ -4,6 +4,16 @@ import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-rout
 
 export default Ember.Route.extend(AuthenticatedRouteMixin, {
     model(params) {
-      return this.store.find('experiment', params.experiment_id);
-    }
+        var self = this;
+        return this.store.find('experiment', params.experiment_id).then(function(experiment) {
+            // When experiment loaded, ensure there are corresponding session models
+            var collId = experiment.get('sessionCollectionId');
+
+            return Ember.RSVP.hash({
+                // The actual return of the model hook: two models, loaded sequentially
+                experiment: experiment,
+                sessions: self.store.findAll(collId),
+            });
+        });
+    },
 });
