+++ app/templates/experiments/info/index.hbs
@@ -1,1 +1,1 @@
-{{ experiment-detail experiment=model editComponents=(action 'editComponents') }}
+{{ experiment-detail experiment=model.experiment sessions=model.sessions editComponents=(action 'editComponents') }}
